pyproject.toml
[project]
name = "openapi-gherkin-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "behave>=1.3.3",
    "google-genai>=1.62.0",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "httpx>=0.27.0",
]
